HSEARCH-3282 Remove support for creating queries from parameters of incorrect type on numeric fields in the migration helper
We used to allow users to pass any value that could be converted to a
numeric value (e.g. Date, Calendar), and we converted that to the
corresponding numeric type (e.g. long). If that type matched the
internal type of the field, everything worked as intended.

This is no longer possible with Search 6, as the internal type of fields
changed in some cases (date/time types in particular) and is now
higher-level (e.g. Date => Instant, not long).

+++ integrationtest/v5migrationhelper/engine/src/test/java/org/hibernate/search/test/engine/numeric/NumericFieldTest.java
@@ -66,8 +66,8 @@ public class NumericFieldTest {
 		// Exact Matching Queries
 		assertExactQuery( "overriddenFieldName", 1 ).as( "Query id exact" ).matchesExactlyIds( 1 );
 		assertExactQuery( "latitude", -10d ).as( "Query double exact" ).matchesExactlyIds( 2 );
-		assertExactQuery( "importance", 12 ).as( "Query short exact" ).matchesExactlyIds( 3 );
-		assertExactQuery( "popularity", 22 ).as( "Query byte exact" ).matchesExactlyIds( 3 );
+		assertExactQuery( "importance", (short)12 ).as( "Query short exact" ).matchesExactlyIds( 3 );
+		assertExactQuery( "popularity", (byte)22 ).as( "Query byte exact" ).matchesExactlyIds( 3 );
 		assertExactQuery( "longitude", -20d ).as( "Query integer exact" ).matchesExactlyIds( 3 );
 		assertExactQuery( "myCounter", 4L ).as( "Query long exact" ).matchesExactlyIds( 4 );
 		assertExactQuery( "strMultiple", 0.1d ).as( "Query multifield exact" ).matchesExactlyIds( 5 );
